Original change: https://android-review.googlesource.com/c/platform/frameworks/support/+/2818598

+++ compose/ui/ui/src/commonMain/kotlin/androidx/compose/ui/platform/ViewConfiguration.kt
@@ -56,7 +56,7 @@ interface ViewConfiguration {
         get() = DpSize(48.dp, 48.dp)
 
     /**
-     * The maximum velocity a fling can start with.
+     * The maximum velocity a fling have at any given time. This value should be in pixels/second.
      */
-    val maximumFlingVelocity: Int get() = Int.MAX_VALUE
+    val maximumFlingVelocity: Float get() = Float.MAX_VALUE
 }
